[2024-06-16] Sprite 공부, window 좌표계, DirectX 좌표계, World 좌표계가 어떻게 변환이 되는지 알아야 한다.

+++ SoloPractice/01_VerySimpleArchitecture/VertexUtil.cpp
@@ -11,7 +11,7 @@ uint16_t AddVertex(BasicVertex* _pVertexList, DWORD _dwMaxVertexCount, DWORD* _p
 	for (DWORD i = 0; i < dwExistVertexCount; i++) {
 		const BasicVertex* pExistVertex = _pVertexList + i;
 		if (!memcmp(pExistVertex, _pVertex, sizeof(BasicVertex))) {
-			dwFoundIndex = i;
+			dwFoundIndex = (uint16_t)i;
 			goto RETURN;
 		}
 	}
